requirements.txt
annotated-types==0.7.0
anyio==4.10.0
cassandra-driver==3.29.2
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.116.1
fastapi-cli==0.0.8
fastapi-cloud-cli==0.1.5
geomet==0.2.1.post1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
kafka-python==2.2.15
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
py4j==0.10.9.9
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyspark==4.0.0
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
rich-toolkit==0.15.0
rignore==0.6.4
sentry-sdk==2.34.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
spark==0.2.1
starlette==0.47.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
watchfiles==1.1.0
websockets==15.0.1
